refactor: journey dto 기존 구조로 통합
