+++ src/main/java/com/waytoearth/controller/v1/Journey/StoryCardController.java
@@ -1,6 +1,6 @@
 package com.waytoearth.controller.v1.Journey;
 
-import com.waytoearth.dto.Journey.response.StoryCardResponse;
+import com.waytoearth.dto.response.journey.StoryCardResponse;
 import com.waytoearth.service.Journey.LandmarkService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
