+++ src/main/java/com/waytoearth/dto/request/journey/JourneyProgressUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.waytoearth.dto.Journey.request;
+package com.waytoearth.dto.request.journey;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
