+++ src/main/java/com/waytoearth/service/Journey/JourneyServiceImpl.java
@@ -1,13 +1,12 @@
 package com.waytoearth.service.Journey;
 
-import com.waytoearth.dto.Journey.request.JourneyStartRequest;
-import com.waytoearth.dto.Journey.request.JourneyProgressUpdateRequest;
-import com.waytoearth.dto.Journey.response.JourneySummaryResponse;
-import com.waytoearth.dto.Journey.response.JourneyProgressResponse;
-import com.waytoearth.dto.Journey.response.JourneyCompletionEstimateResponse;
-import com.waytoearth.dto.Journey.response.LandmarkSummaryResponse;
+import com.waytoearth.dto.request.journey.JourneyStartRequest;
+import com.waytoearth.dto.request.journey.JourneyProgressUpdateRequest;
+import com.waytoearth.dto.response.journey.JourneySummaryResponse;
+import com.waytoearth.dto.response.journey.JourneyProgressResponse;
+import com.waytoearth.dto.response.journey.JourneyCompletionEstimateResponse;
+import com.waytoearth.dto.response.journey.LandmarkSummaryResponse;
 import com.waytoearth.entity.Journey.JourneyEntity;
-import com.waytoearth.entity.Journey.LandmarkEntity;
 import com.waytoearth.entity.Journey.UserJourneyProgressEntity;
 import com.waytoearth.entity.User.User;
 import com.waytoearth.repository.Journey.JourneyRepository;
