+++ src/main/java/com/waytoearth/service/Journey/LandmarkService.java
@@ -1,8 +1,8 @@
 package com.waytoearth.service.Journey;
 
-import com.waytoearth.dto.Journey.response.LandmarkDetailResponse;
-import com.waytoearth.dto.Journey.response.LandmarkSummaryResponse;
-import com.waytoearth.dto.Journey.response.StoryCardResponse;
+import com.waytoearth.dto.response.journey.LandmarkDetailResponse;
+import com.waytoearth.dto.response.journey.LandmarkSummaryResponse;
+import com.waytoearth.dto.response.journey.StoryCardResponse;
 import com.waytoearth.entity.Journey.StoryCardEntity;
 
 import java.util.List;
